+++ vocalocatorssl/src/location_embed.py
